refactor: 역할별로 패키지 조정
