+++ README.md
@@ -52,7 +52,7 @@
 - [x] 로또 발매기를 인터페이스로 추상화하고 구체화해서 사용하며 원시값을 사용하지 않는다.
 - [x] 비논리 오류 예외를 컨트롤러에서 null로 처리한다.
 - [x] 모호한 함수명을 명확화한다.
-- [ ] 패키지명이 의도에 맞는지, 역할별로 모아져 있는지 다시 확인한다.
+- [x] 패키지명이 의도에 맞는지, 역할별로 모아져 있는지 다시 확인한다.
 
 ## 동작 과정
 1. 컨트롤러가 게임 실행 및 진행을 관리한다
