+++ src/main/kotlin/lotto/domain/model/LottoPayInfo.kt
@@ -1,4 +1,4 @@
-package lotto.domain.value
+package lotto.domain.model
 
 import lotto.constants.LottoConstants.LOTTO_PRICE
