+++ src/main/kotlin/lotto/domain/model/lottomachine/LottoMachine.kt
@@ -1,8 +1,8 @@
 package lotto.domain.model.lottomachine
 
 import lotto.domain.model.Lotto
-import lotto.domain.value.LottoNumber
-import lotto.domain.value.LottoPayInfo
+import lotto.domain.model.LottoPayInfo
+import lotto.domain.valueobject.LottoNumber
 
 interface LottoMachine {
     fun generateLottoBundle(
