+++ src/main/kotlin/lotto/domain/valueobject/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto.domain.value
+package lotto.domain.valueobject
 
 @JvmInline
 value class LottoNumber(
