+++ src/test/kotlin/lottotest/domain/model/lottomachine/AutoLottoMachineTest.kt
@@ -1,8 +1,8 @@
-package lotto.domain
+package lottotest.domain.model.lottomachine
 
 import lotto.domain.model.Lotto
+import lotto.domain.model.LottoPayInfo
 import lotto.domain.model.lottomachine.AutoLottoMachine
-import lotto.domain.value.LottoPayInfo
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
